requirements.txt
# AMPR: Adaptive Multimodal Protein Representation
# Python 3.12.x compatible versions (Colab + Kaggle pinned)

# Core ML
torch==2.3.1
transformers==4.41.2
tensorflow==2.16.1

# Graph processing
dgl==2.1.0

# Bio data formats
obonet==1.0.0
biopython==1.84

# Data + utilities
numpy==1.26.4
scipy==1.14.0
pandas==2.2.0
scikit-learn==1.5.0

# Config + logging
pyyaml==6.0.1
tqdm==4.66.4

# Optional: visualization (for notebooks)
matplotlib==3.8.4
seaborn==0.13.2

# Optional: testing
pytest==7.4.4
pytest-cov==4.1.0
